+++ packages/plugins/didkit-plugin-node/src/types.ts
@@ -1,0 +1,1 @@
+export type { DIDKitPlugin, DidMethod, ProofOptions, InputMetadata } from '@learncard/didkit-plugin';